api/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Pillow==10.4.0
piexif==1.1.3
numpy==1.26.4